requirements.txt
Flask==3.0.0
RPi.GPIO==0.7.1
Werkzeug==3.0.1
pygame==2.5.2
numpy==1.24.3

